+++ cmd/program/program.go
@@ -3,19 +3,15 @@
 package program
 
 import (
-	"bytes"
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"text/template"
 
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/melkeydev/go-blueprint/cmd/flags"
-	tpl "github.com/melkeydev/go-blueprint/cmd/template"
 	"github.com/melkeydev/go-blueprint/cmd/template/advanced"
 	"github.com/melkeydev/go-blueprint/cmd/template/backend"
 	"github.com/melkeydev/go-blueprint/cmd/template/dbdriver"
@@ -128,22 +124,6 @@ const (
 	internalDatabasePath = "internal/database"
 	gitHubActionPath     = ".github/workflows"
 )
-
-// CheckOs checks Operation system and generates MakeFile and `go build` command
-// Based on Project.Unixbase
-func (p *Project) CheckOS() {
-	p.OSCheck = make(map[string]bool)
-
-	if runtime.GOOS != "windows" {
-		p.OSCheck["UnixBased"] = true
-	}
-	if runtime.GOOS == "linux" {
-		p.OSCheck["linux"] = true
-	}
-	if runtime.GOOS == "darwin" {
-		p.OSCheck["darwin"] = true
-	}
-}
 
 // ExitCLI checks if the Project has been exited, and closes
 // out of the CLI if it has
@@ -723,281 +703,6 @@ func (p *Project) CreateMainFile() error {
 				return err
 			}
 		}
-	}
-	return nil
-}
-
-// CreatePath creates the given directory in the projectPath
-func (p *Project) CreatePath(pathToCreate string, projectPath string) error {
-	path := filepath.Join(projectPath, pathToCreate)
-	if _, err := os.Stat(path); os.IsNotExist(err) {
-		err := os.MkdirAll(path, 0o751)
-		if err != nil {
-			log.Printf("Error creating directory %v\n", err)
-			return err
-		}
-	}
-
-	return nil
-}
-
-// CreateFileWithInjection creates the given file at the
-// project path, and injects the appropriate template
-func (p *Project) CreateFileWithInjection(pathToCreate string, projectPath string, fileName string, methodName string) error {
-	createdFile, err := os.Create(filepath.Join(projectPath, pathToCreate, fileName))
-	if err != nil {
-		return err
-	}
-
-	defer createdFile.Close()
-
-	switch methodName {
-	case "main":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(p.BackendMap[p.ProjectType].templater.Main())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "server":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(p.BackendMap[p.ProjectType].templater.Server())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "routes":
-		routeFileBytes := p.BackendMap[p.ProjectType].templater.Routes()
-		createdTemplate := template.Must(template.New(fileName).Parse(string(routeFileBytes)))
-		err = createdTemplate.Execute(createdFile, p)
-	case "releaser":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(advanced.Releaser())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "go-test":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(advanced.Test())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "releaser-config":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(advanced.ReleaserConfig())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "database":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(p.DBDriverMap[p.DBDriver].templater.Service())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "db-docker":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(p.DockerMap[p.Docker].templater.Docker())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "integration-tests":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(p.DBDriverMap[p.DBDriver].templater.Tests())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "tests":
-		createdTemplate := template.Must(template.New(fileName).Parse(string(p.BackendMap[p.ProjectType].templater.TestHandler())))
-		err = createdTemplate.Execute(createdFile, p)
-	case "env":
-		if p.DBDriver != "none" {
-
-			envBytes := [][]byte{
-				tpl.GlobalEnvTemplate(),
-				p.DBDriverMap[p.DBDriver].templater.Env(),
-			}
-			createdTemplate := template.Must(template.New(fileName).Parse(string(bytes.Join(envBytes, []byte("\n")))))
-			err = createdTemplate.Execute(createdFile, p)
-
-		} else {
-			createdTemplate := template.Must(template.New(fileName).Parse(string(tpl.GlobalEnvTemplate())))
-			err = createdTemplate.Execute(createdFile, p)
-		}
-	}
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (p *Project) CreateViteReactProject(projectPath string) error {
-	if err := checkNpmInstalled(); err != nil {
-		return err
-	}
-
-	originalDir, err := os.Getwd()
-	if err != nil {
-		return fmt.Errorf("failed to get current directory: %w", err)
-	}
-	defer func() {
-		if err := os.Chdir(originalDir); err != nil {
-			fmt.Fprintf(os.Stderr, "failed to change back to original directory: %v\n", err)
-		}
-	}()
-
-	// change into the project directory to run vite command
-	err = os.Chdir(projectPath)
-	if err != nil {
-		fmt.Println("failed to change into project directory: %w", err)
-	}
-
-	// the interactive vite command will not work as we can't interact with it
-	fmt.Println("Installing create-vite (using cache if available)...")
-	cmd := exec.Command("npm", "create", "vite@latest", "frontend", "--",
-		"--template", "react-ts",
-		"--prefer-offline",
-		"--no-fund")
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to use create-vite: %w", err)
-	}
-
-	frontendPath := filepath.Join(projectPath, "frontend")
-	if err := os.MkdirAll(frontendPath, 0755); err != nil {
-		return fmt.Errorf("failed to create frontend directory: %w", err)
-	}
-
-	if err := os.Chdir(frontendPath); err != nil {
-		return fmt.Errorf("failed to change to frontend directory: %w", err)
-	}
-
-	srcDir := filepath.Join(frontendPath, "src")
-	if err := os.MkdirAll(srcDir, 0755); err != nil {
-		return fmt.Errorf("failed to create src directory: %w", err)
-	}
-
-	if err := os.WriteFile(filepath.Join(srcDir, "App.tsx"), frontend.ReactAppfile(), 0644); err != nil {
-		return fmt.Errorf("failed to write App.tsx template: %w", err)
-	}
-
-	// Create the global `.env` file from the template
-	err = p.CreateFileWithInjection("", projectPath, ".env", "env")
-	if err != nil {
-		return fmt.Errorf("failed to create global .env file: %w", err)
-	}
-
-	// Read from the global `.env` file and create the frontend-specific `.env`
-	globalEnvPath := filepath.Join(projectPath, ".env")
-	vitePort := "8080" // Default fallback
-
-	// Read the global .env file
-	if data, err := os.ReadFile(globalEnvPath); err == nil {
-		lines := strings.Split(string(data), "\n")
-		for _, line := range lines {
-			if strings.HasPrefix(line, "PORT=") {
-				vitePort = strings.SplitN(line, "=", 2)[1] // Get the backend port value
-				break
-			}
-		}
-	}
-
-	// Use a template to generate the frontend .env file
-	frontendEnvContent := fmt.Sprintf("VITE_PORT=%s\n", vitePort)
-	if err := os.WriteFile(filepath.Join(frontendPath, ".env"), []byte(frontendEnvContent), 0644); err != nil {
-		return fmt.Errorf("failed to create frontend .env file: %w", err)
-	}
-
-	// Handle Tailwind configuration if selected
-	if p.FrontendOptions[string(flags.Tailwind)] && p.FrontendFramework == flags.React {
-		fmt.Println("Installing Tailwind dependencies (using cache if available)...")
-		cmd := exec.Command("npm", "install",
-			"--prefer-offline",
-			"--no-fund",
-			"tailwindcss", "postcss", "autoprefixer")
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to install Tailwind: %w", err)
-		}
-
-		fmt.Println("Initializing Tailwind...")
-		cmd = exec.Command("npx", "--prefer-offline", "tailwindcss", "init", "-p")
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to initialize Tailwind: %w", err)
-		}
-
-		// use the tailwind config file
-		err = os.WriteFile("tailwind.config.js", frontend.ReactTailwindConfigTemplate(), 0644)
-		if err != nil {
-			return fmt.Errorf("failed to write tailwind config: %w", err)
-		}
-
-		srcDir := filepath.Join(frontendPath, "src")
-		if err := os.MkdirAll(srcDir, 0755); err != nil {
-			return fmt.Errorf("failed to create src directory: %w", err)
-		}
-
-		err = os.WriteFile(filepath.Join(srcDir, "index.css"), frontend.InputCssTemplateReact(), 0644)
-		if err != nil {
-			return fmt.Errorf("failed to update index.css: %w", err)
-		}
-
-		if err := os.WriteFile(filepath.Join(srcDir, "App.tsx"), frontend.ReactTailwindAppfile(), 0644); err != nil {
-			return fmt.Errorf("failed to write App.tsx template: %w", err)
-		}
-
-		if err := os.Remove(filepath.Join(srcDir, "App.css")); err != nil {
-			// Don't return error if file doesn't exist
-			if !os.IsNotExist(err) {
-				return fmt.Errorf("failed to remove App.css: %w", err)
-			}
-		}
-	}
-
-	return nil
-}
-func (p *Project) CreateHtmxTemplates() {
-	routesPlaceHolder := ""
-	importsPlaceHolder := ""
-	if p.FrontendFramework == flags.Htmx {
-		routesPlaceHolder += string(p.BackendMap[p.ProjectType].templater.HtmxTemplRoutes())
-		importsPlaceHolder += string(p.BackendMap[p.ProjectType].templater.HtmxTemplImports())
-	}
-
-	routeTmpl, err := template.New("routes").Parse(routesPlaceHolder)
-	if err != nil {
-		log.Fatal(err)
-	}
-	importTmpl, err := template.New("imports").Parse(importsPlaceHolder)
-	if err != nil {
-		log.Fatal(err)
-	}
-	var routeBuffer bytes.Buffer
-	var importBuffer bytes.Buffer
-	err = routeTmpl.Execute(&routeBuffer, p)
-	if err != nil {
-		log.Fatal(err)
-	}
-	err = importTmpl.Execute(&importBuffer, p)
-	if err != nil {
-		log.Fatal(err)
-	}
-	p.FrontendTemplates.TemplateRoutes = routeBuffer.String()
-	p.FrontendTemplates.TemplateImports = importBuffer.String()
-}
-
-func (p *Project) CreateWebsocketImports(appDir string) {
-	websocketDependency := []string{"github.com/coder/websocket"}
-	if p.ProjectType == flags.Fiber {
-		websocketDependency = []string{"github.com/gofiber/contrib/websocket"}
-	}
-
-	// Websockets require a different package depending on what backend is
-	// choosen. The application calls go mod tidy at the end so we don't
-	// have to here
-	err := utils.GoGetPackage(appDir, websocketDependency)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	importsPlaceHolder := string(p.BackendMap[p.ProjectType].templater.WebsocketImports())
-
-	importTmpl, err := template.New("imports").Parse(importsPlaceHolder)
-	if err != nil {
-		log.Fatalf("CreateWebsocketImports failed to create template: %v", err)
-	}
-	var importBuffer bytes.Buffer
-	err = importTmpl.Execute(&importBuffer, p)
-	if err != nil {
-		log.Fatalf("CreateWebsocketImports failed write template: %v", err)
-	}
-	newImports := strings.Join([]string{string(p.AdvancedTemplates.TemplateImports), importBuffer.String()}, "\n")
-	p.AdvancedTemplates.TemplateImports = newImports
-}
-
-func checkNpmInstalled() error {
-	cmd := exec.Command("npm", "--version")
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("npm is not installed: %w", err)
 	}
 	return nil
 }
